[FIX] : article-fix / #91
아티클 티켓 받아오는 시점 변경

+++ app/src/main/java/com/indipage/presentation/articledetail/ArticleDetailViewModel.kt
@@ -1,5 +1,6 @@
 package com.indipage.presentation.articledetail
 
+import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
@@ -58,8 +59,8 @@ class ArticleDetailViewModel @Inject constructor(
         }.onFailure { Timber.d(it.message.toString()) }
     }
 
-    fun getTicketReceiveCheck(spaceId: Long) = viewModelScope.launch {
-        apiRepository.getTicketReceiveCheck(spaceId).onSuccess {
+    fun getTicketReceiveCheck(spaceId: Int) = viewModelScope.launch {
+        apiRepository.getTicketReceiveCheck(spaceId.toLong()).onSuccess {
             _ticketReceiveCheckData.value = it
             Timber.d("Success")
         }.onFailure { Timber.d(it.message.toString()) }
